external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/canvas/viewport-unchanged-upon-resize-expected.txt

Verifies that GL viewport does not change when canvas is resized

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getError was expected value: NO_ERROR : No GL errors before resizing the canvas
PASS err is gl.NO_ERROR
PASS Viewport correctly did not change size during canvas resize
PASS successfullyParsed is true

TEST COMPLETE
